Duc trung back end

+++ src/main/java/com/nhom7/ecommercebackend/model/OrderDetail.java
@@ -3,6 +3,8 @@ package com.nhom7.ecommercebackend.model;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "order_details")
 @Getter
@@ -17,7 +19,6 @@ public class OrderDetail {
 
     @ManyToOne
     @JoinColumn(name = "order_id")
-//    @JsonBackReference
     private Order order;
 
     @ManyToOne
@@ -28,7 +29,6 @@ public class OrderDetail {
     private int numberOfProducts;
 
     @Column(name = "total_money", nullable = false)
-    private Float totalMoney;
-
+    private BigDecimal totalMoney;
 
 }
